requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
requests
gradio-client
Pillow
opencv-python-headless
gunicorn
